Make attelo.decoding.astar test work again
The code had changed to use an EDU object rather than strings,
and we didn't notice because of dynamic typing.

We simulate these EDU objects instead.

Note that this code uses deepcopy on its discourse state,
which is potentially problematic because EDUs also get deepcopied,
but the search uses dictionary lookup, list membership, etc.

As a fix, it should suffice to make the EDU deepcopy just return
the object itself, but in order for that to be a reasonable thing
to do, we should enforce immutability on the EDU members.