Bugfix: return rels from prob dist in MST decoder
Previously, the MST decoder was systematically returning the
same label for all links in the same graph. This is because
the bit that returns labels wasn't actually implemented, and
it was instead was grabbing the labels from a value in a for
loop that should have fallen out of scope.

The oversight went unnoticed due to Python silliness:
https://mail.python.org/pipermail/python-ideas/2008-October/002109.html