Continue astar arguments refactor
The hope is increased modularity in our arguments processing. Prior to
this strand of work, any option we wanted to introduce to A* would have
to be splattered throughout the code.

We still have to define them in a a couple of places:

* astar args
* shared state
* argument parsing

But maybe this will help us to get a better handle on it