Improve mechanism for passing arguments to attelo
This refactor should buy us some extra robustness with respect
to evolution within attelo. The reason we were trying to
define the namespace objects in the first place was so that we
could bypass parts of attelo commands that would be wasefully
repeated (over folds, parameters, etc). But a problem that came
up when we tried to define them by hand was that we had to define
*all* of the parameters, even those we weren't interested in.
Worse, whenever attelo evolved and grew a parameter or two, the
evaluation framework would break.

Our new simpler approach just asks argparse to parse a simulated command
line and create the object for us.  (As a bonus, this would also allow
us to generate the arguments we'd need to call attelo by hand rather
than invoking the calls through the API)