Raise exception when user asks for too many folds
Closes #7

Alternatively, we could have silently generated the min
of the two, but I suspect this would have knock-on
consequences for test harnesses that may not themselves
have anticipated the problem. The worst would be if it
silently does the wrong thing.

At least this way we can more quickly identify the problem