Fix/refactor single instance score retrieval
This was subtly broken for the corner case where the classifier only
ever returns 'True' or 'False', but it didn't crash because it was
hidden by #12

We would try to retrieve the probability for value 1, on the assumption
that 0 was the index for 'False', and 1 the index for 'True'. This
*would* crash on lookup failure except that relation labels were also
included in the range of possible values (the variable conflation issue
in #12), and instead we were picking out the probability associated with
one of the relation labels... which in practice probably gave us the
same answer anyway.

But it's still wrong.