Fix regression introduced by 424f4fd0
This was causing us to always return probabilities of 0,
which in turn caused us to get zero scores everywhere.

If I understand correctly, Orange distributions behave like dictionaries
for indexing but not for things like `in` keyword.

This is a sort of type error, I think, and if so, SIGH