MAINT Refactor learning
Implemented in attelo.learning.control and re-exported in
attelo.learning

I don't really like these control modules floating around,
or the fact that I don't have a distinction between
decoders and decoding, nor learners and learning. Part of
the issue is that I'm still a bit nervous about putting
substantial bits of code in the __init__ module of any
subpackage.