FIX confusion matrix mutation bug
The logic was supposed to associate each new configuration with
a zero matrix to start with.

But I was initialising everything to the same "zero" object and
then incrementing it in (confusion[key] += ...), which meant
that everything  we added to the matrices was all mixed in one
big pot.

Ugh, still thinking in immutable-lanugage terms.