MAINT move utility functions out of cmd
This patch replicates some work in irit-stac.
There it was motivated by us wanting to reuse these
functions in the standalone pipeline.

The main idea here is that we should try and thin
out the cmd.evaluate smarts as much as possible
and move the core business logic into dedicated
modules.

In the long run we want to have more sharing between
the two harnesses