MAINT call decoder.transform not decode
This allows us to use parsing pipelines in place of decoders.
For example, we could make a mini pipeline out of a pruner
and a decoder. The intention would still be to use a decoder
and not a full-fledged pipeline... but if you really had cause
to, you could also insert a pipeline in place. Mind you don't
create some kind of infinite loop.