ENH improve learn/decode parallelism
I just noticed that it seems possible to use joblib parallel
in a sort of producer/consumer pattern: we don't need to have
all our jobs ready in one go

So if I understand correctly, this means we can have a generator
expression that yields decoder jobs as soon as the parser for
them has been fit.

Outcome:

1. No more waiting for all the learners to complete before we
   decode; start decoding as soon as the relevant learner is
   done
2. But still apply parallelism across decoders for multiple
   configurations

So if all learners are fast, that's great, we get to run lots
of jobs in parallel.  If some learners are slow, that's hopefully
still OK because we can still work on decoding while they're
crunching away.

I hope this means less dead space.  More cores humming away in
parallel