FIX ignore cross-sentential links in intrasentential decoding
We were trying to be concise and picking out sentences by the
subgrouping by the second EDU only (I think this must have been to avoid
looking at the fake root), but this completely ignores the existence of
cross-sentential links. In fact any cross-sentential edge into the
sentence is included for consideration.