FIX ugh work around corner case triggered by test
If we have too few edges, we can easily trigger the case where the
solver only sees one class. Here the errors came fram a lack of
intersentential links to filter on.

It's slightly concerning because you can imagine situations where
this sort of thing may come up in the wild.